make EE search APIs CE
