+++ backend/tests/integration/common_utils/managers/document_search.py
@@ -1,10 +1,10 @@
 import requests
 
-from ee.onyx.server.query_and_chat.models import DocumentSearchRequest
 from onyx.context.search.enums import LLMEvaluationType
 from onyx.context.search.enums import SearchType
 from onyx.context.search.models import RetrievalDetails
 from onyx.context.search.models import SavedSearchDocWithContent
+from onyx.server.query_and_chat.models import DocumentSearchRequest
 from tests.integration.common_utils.constants import API_SERVER_URL
 from tests.integration.common_utils.constants import GENERAL_HEADERS
 from tests.integration.common_utils.test_models import DATestUser
