+++ backend/tests/regression/answer_quality/api_utils.py
@@ -1,7 +1,6 @@
 import requests
 from retry import retry
 
-from ee.onyx.server.query_and_chat.models import OneShotQARequest
 from onyx.chat.models import ThreadMessage
 from onyx.configs.constants import DocumentSource
 from onyx.configs.constants import MessageType
@@ -11,6 +10,7 @@ from onyx.context.search.models import IndexFilters
 from onyx.context.search.models import RetrievalDetails
 from onyx.db.enums import IndexingStatus
 from onyx.server.documents.models import ConnectorBase
+from onyx.server.query_and_chat.models import OneShotQARequest
 from tests.regression.answer_quality.cli_utils import get_api_server_host_port
 
 GENERAL_HEADERS = {"Content-Type": "application/json"}
